+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `get cc` returns the credentials of a cluster claim [Retrieve the credentials of a given cluster](https://github.com/open-cluster-management/cm-cli/issues/39)
 - `scale cp <clusterpool_name> --size <size>`
 - `get cps [<clusterpoolhost>|-A]` to get the list of clusterpools
+- `attach cc <clusterclaim>` [Needs for a cm attach clusterclaim --cluster <cluster_name>](https://github.com/open-cluster-management/cm-cli/issues/51)
 ### Additions
 - Add the command `delete token` from [clusteradm](https://github.com/open-cluster-management-io/clusteradm)
 ### Breacking changes